moves web routes into separate controller
